% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.strata.R
\docType{data}
\name{cat.strata}
\alias{cat.strata}
\title{MW Catchment strata}
\format{A data frame with nested catchment tables (Pfaffstetter coding and Nick's strata)
\describe{
\item{site}{}
\item{segmentno}{}
\item{NCB_CODE}{}
\item{PFAFNO}{}
\item{LEVEL1}{}
\item{DIVISION}{}
\item{LEVEL2}{}
\item{LEVEL2NAME}{}
\item{PFAFL3}{}
\item{PFAFL4}{}
\item{PFAFL5}{}
\item{PFAFL6}{}
\item{PFAFL7}{}
\item{PFAFL8}{}
\item{PFAFL9}{}
\item{PFAFL10}{}
\item{PFAFL11}{}
\item{PFAFL12}{}
\item{PFAFL13}{}
\item{PFAFL14}{}
\item{PFAFL15}{}
}}
\source{
Nick Bond
}
\usage{
cat.strata
}
\description{
cat.strata
}
\keyword{datasets}

