% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dam.catchments.R
\docType{data}
\name{dam.catchments}
\alias{dam.catchments}
\title{Table identifying catchments in or above onstream storages.}
\format{A data frame with 21 observations and 6 variables
\describe{
\item{closed.cat}{}
\item{nextds}{}
\item{dam}{}
\item{upstream.excluded.from.carea}{}
\item{propn.subc.us.of.dam}{}
\item{comment}{}
}}
\source{
Nick Bond
}
\usage{
dam.catchments
}
\description{
dam.catchments
}
\keyword{datasets}

