% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelled.temps.R
\docType{data}
\name{modelled.temps}
\alias{modelled.temps}
\title{Modelled stream temperature data.}
\format{A data frame with 12319 observations and 10 air and water temperature variables
\describe{
\item{site}{}
\item{mxSumAirTm}{}
\item{mnAnnAirTm}{}
\item{mnAnnStreamTm}{}
\item{NHImnAnnStreamTm}{}
\item{movamxStreamTm}{}
\item{NHImovamxStreamTm}{}
\item{CCmnAnnStreamTm}{}
\item{CCmovamxStreamTm}{}
\item{CCNHImnAnnStreamTm}{}
\item{CCNHImovamxStreamTm}{}
}}
\source{
Nick Bond
}
\usage{
modelled.temps
}
\description{
Modelled stream temperature data derived from BRT models built using land-use, hydrology and climate
variables. Unpublished results (Walsh and Bond, 2015).
}
\keyword{datasets}

