% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwdata-package.r
\docType{package}
\name{mwdata}
\alias{mwdata}
\alias{mwdata-package}
\title{mwdata: A set of physiographic, climate and biological datasets for the Port Philip region, southeastern
Australia.}
\description{
mwdata: A set of physiographic, climate and biological datasets for the Port Philip region, southeastern
Australia.
}

