% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.codes.R
\docType{data}
\name{species.codes}
\alias{species.codes}
\title{Species names and abbreviations}
\format{A data frame with 150 species names and 11 summary variables.
\describe{
\item{COMMON.NAME}{}
\item{SCIENTIFIC.NAME}{}
\item{SPECIES.CODE}{}
\item{SHORT.SPECIES.CODE}{}
\item{CATEGORY}{}
\item{TAXON}{}
\item{MIGRATORY}{}
\item{DSE.ADVISORY.LIST}{}
\item{FFG.STATUS}{}
\item{EPBC.STATUS}{}
\item{EXOTIC}{}
}}
\source{
Nick Bond
}
\usage{
species.codes
}
\description{
Table of fish species names and abbreviation codes based on scientific names. Also includes conservation status
whether species are native or exotic. Table data is incomplete and includes many marine species and some crustaceans
}
\details{
species.codes
}
\keyword{datasets}

