% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/werg.nw.R
\docType{data}
\name{werg.nw}
\alias{werg.nw}
\title{MW Stream Network data (WERG Network)}
\format{A data frame with attributes for 15901 stream segments, including above and
below ground pipes, open channels, and natural waterways.
\describe{
\item{site}{}
\item{nextds}{}
\item{buffL30W30000}{}
\item{buffL30W1000}{}
}}
\source{
Nick Bond
}
\usage{
werg.nw
}
\description{
werg.nw
}
\keyword{datasets}

